+++ trunk/src/se/umu/cs/ldbn/client/core/AttributeSetIterator.java
@@ -3,6 +3,6 @@ package se.umu.cs.ldbn.client.core;
 import java.util.Iterator;
 
 public interface AttributeSetIterator extends Iterator<String> {
-	public long nextAttIndex();
+	public int nextAttIndex();
 	public String nextAttName();
 }
